[#20] - add pagination and fix api calls twice
